+++ dud.cpp
@@ -10,7 +10,7 @@
 
 using namespace Starlia;
 
-Dud::Dud(Coordinate2d position, Coordinate2d velocity, double radius, Color3d color)
+Dud::Dud(Coordinate2d position, Coordinate2d velocity, double radius, Color3f color)
 	: position(position), velocity(velocity), radius(radius), color(color)
 {
 }
@@ -33,20 +33,20 @@ bool Dud::recalc()
 	}
 	else if (ground->collides(position))
 	{
-		layer[1]->registerObject(new P_engine(Coordinate2d(position.x, ground->getYofX(position.x)),PARTICLES,TIME,SPEED,Color3d(1,0.3,0),P_engine::CHAMPAGNE), changePlayer);
+		layer[1]->registerObject(new P_engine(Coordinate2d(position.x, ground->getYofX(position.x)),PARTICLES,TIME,SPEED,Color3f(1,0.3,0),P_engine::CHAMPAGNE), changePlayer);
 		StarSound::playSound("explosion");
 		return false;
 	}
 	else if (player[0]->collides(position))
 	{
-		layer[1]->registerObject(new P_engine(position,PARTICLES,TIME,SPEED,Color3d(1,0,0),P_engine::CHAMPAGNE), reInit);
+		layer[1]->registerObject(new P_engine(position,PARTICLES,TIME,SPEED,Color3f(1,0,0),P_engine::CHAMPAGNE), reInit);
 		++score[1];
 		StarSound::playSound("explosion");
 		return false;
 	}
 	else if (player[1]->collides(position))
 	{
-		layer[1]->registerObject(new P_engine(position,PARTICLES,TIME,SPEED,Color3d(1,0,0),P_engine::CHAMPAGNE), reInit);
+		layer[1]->registerObject(new P_engine(position,PARTICLES,TIME,SPEED,Color3f(1,0,0),P_engine::CHAMPAGNE), reInit);
 		++score[0];
 		StarSound::playSound("explosion");
 		return false;
