+++ ground.h
@@ -12,10 +12,10 @@ class Ground : public StarObject
 {
 	private:
 		double points[GR_SIZE]; 	// the ground set points
-		Color3d color;
+		Color3f color;
 
 	public:
-		Ground(Color3d);
+		Ground(Color3f);
 		double getYofX(double x);  	// returns Y of given X
 		void draw(); 			// paints the ground, when initialised
 		void indent(double x);		// intents the ground in x
